# docs/swagger.yaml
definitions:
  executor.ExeLog:
    properties:
      line:
        description: 行内容
        type: string
      output_at:
        description: 输出时间
        type: string
      type:
        description: 内容类型，0标准输出，1标准错误输出
        type: integer
    type: object
  request.Task:
    properties:
      bash_content:
        type: string
      description:
        type: string
      name:
        type: string
      run_on:
        type: string
    type: object
  response.Run:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      exit_code:
        type: integer
      id:
        type: integer
      logs:
        items:
          $ref: '#/definitions/executor.ExeLog'
        type: array
      name:
        type: string
      start_time:
        type: string
      status:
        type: integer
      time_elapse_second:
        type: integer
    type: object
  response.Task:
    properties:
      bash:
        type: string
      cron:
        type: string
      id:
        type: integer
      last_run:
        $ref: '#/definitions/response.Run'
      name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/tasks:
    get:
      description: get task list
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Task'
            type: array
      summary: get task list
      tags:
      - Task
    post:
      description: create a task
      parameters:
      - description: Bearer 31a165ba1be6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: JSON数据
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.Task'
      summary: creates a task
      tags:
      - Task
...
  /api/v1/tasks/:task_id:
    put:
      description: update a task
      parameters:
      - description: Bearer 31a165ba1be6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: JSON数据
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.Task'
      summary: updates a task
      tags:
      - Task
...
swagger: "2.0"